remove unused code and rename file
